backend/requirements.txt
fastapi[standard]>=0.110.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
passlib[bcrypt]>=1.7.4
bcrypt<5.0
python-jose[cryptography]>=3.3.0
cloudinary>=1.36.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
apscheduler>=3.10.0
firebase-admin>=6.5.0
